Implement order loading

+++ internal/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID       int64  `json:"id,omitempty"`
+	ID       int    `json:"id,omitempty"`
 	Login    string `json:"login"`
 	Password string `json:"password"`
 }
